+++ backend/quiz-server/mysql-init/schema.sql
@@ -3,8 +3,8 @@ CREATE DATABASE IF NOT EXISTS quiz_platform;
 USE quiz_platform;
 
 CREATE TABLE `Quiz` (
-    `quiz_id` INT NOT NULL AUTO_INCREMENT,
-    `crawling_id` INT NOT NULL,
+    `quiz_id` BIGINT NOT NULL AUTO_INCREMENT,
+    `crawling_id` BIGINT NOT NULL,
     `content` TEXT NOT NULL,
     `quiz_content` TEXT NOT NULL,
     `quiz_answer` BOOLEAN NOT NULL,
