+++ backend/summary-server/mysql-init/schema.sql
@@ -3,8 +3,8 @@ CREATE DATABASE IF NOT EXISTS summary_platform;
 USE summary_platform;
 
 CREATE TABLE `Summary` (
-    `summary_id` INT NOT NULL AUTO_INCREMENT,
-    `article_id` INT NOT NULL,
+    `summary_id` BIGINT NOT NULL AUTO_INCREMENT,
+    `article_id` BIGINT NOT NULL,
     `content` TEXT NOT NULL,
     `quiz_content` TEXT NOT NULL,
     `quiz_answer` BOOLEAN NOT NULL,
